Fix V4 and MCTS engine compilation errors
Fixed all Position API compatibility issues across V4 search, MCTS search, tournament, PGN, and evaluation modules:

- Corrected make_move() usage pattern (clone position first, then mutate)
- Fixed is_in_check() calls to use is_check()
- Changed pos.hash() field access to pos.hash
- Fixed piece_at() parameter types (u8 instead of usize)
- Added Hash derive to EngineType for HashMap usage
- Fixed TTEntry store() call signature
- Converted MoveList to Vec for move ordering
- Temporarily disabled null move pruning and repetition detection (requires Position API additions)
- Disabled multi-cut pruning temporarily (API complexity)

All engines now compile successfully and UCI engine verified working (depth 3 search from startpos returns b1c3).

Test results: 0 errors, 25 warnings (all unused imports/variables)

+++ backend/src/engine/evaluation_v3.rs
@@ -55,7 +55,7 @@ impl EvaluatorV3 {
         let mut black_bishops = 0;
 
         for sq in 0..64 {
-            if let Some((piece_type, color)) = pos.piece_at(sq) {
+            if let Some((piece_type, color)) = pos.piece_at(sq as u8) {
                 let base_value = match piece_type {
                     PieceType::Pawn => PAWN_VALUE,
                     PieceType::Knight => KNIGHT_VALUE,
